requirements.txt
amqp==2.2.2
Babel==2.4.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.4.17
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.3
django-allauth==0.32.0
django-appconf==1.0.2
django-braces==1.11.0
django-compressor==2.1.1
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-easy-select2==1.3.4
django-extensions==1.8.0
django-extra-views==0.9.0
django-filter==1.1.0
django-mptt==0.8.7
django-phonenumber-field==1.3.0
django-redis==4.8.0
django-suit==0.2.25
django-tables2==1.10.0
docutils==0.14
idna==2.5
kombu==4.1.0
oauthlib==2.0.2
olefile==0.44
phonenumberslite==8.5.2
Pillow==4.2.0
psycopg2==2.7.1
python-dateutil==2.6.0
python-openid==2.2.5
python3-openid==3.1.0
pytz==2017.2
rcssmin==1.0.6
redis==2.10.6
requests==2.18.1
requests-oauthlib==0.8.0
rjsmin==1.0.12
six==1.10.0
sqlparse==0.2.3
statistics==1.0.3.5
tablib==0.12.1
urllib3==1.21.1
vine==1.1.4
whitenoise==3.3.0
